feat(input): Add class definition for PETLINKStream with constructor and destructor
Found out that the data format of the listmode data is published by Siemens here:
https://www.siemens-healthineers.com/en-us/molecular-imaging/petlink-documents

Will now build the parser after that and wrap everything into proper classes to
have a more easily readable codebase.

The PETLINKStream inherits from std::ifstream as it seems to me that this is easier
than maintaining an instance of the ifstream as an attribute.