fix(event/tag): Use value type properties for event and tag
When switching to the shared pointer the union stopped working because
of the "nontrivial constructor" of shared_ptr.
I have no clue what that means but now just removed this as a whole
and use the struct with two attributes.

I think the benefit of using a union would be that no memory is allocated for both
objects, however, if the following logic extracts the event or tag and stores it in
appropriate form, the increased size of EventOrTag should not be of significance.