fix(petlinkstream): Middle in binary search must be multiple of four
The positions are given in bytes, however, one word has four bytes
and therefore we can end up in the middle of a word if the middle
position is no multiple of four.

Rounding it down guarantees that we always end up on the beginning
of an event or tag.