feat(PETLINKStream): Add iterator to petlinkstream
Trying this out as it makes iterating over the elements in the stream
much easier. Not yet fully sure when exactly the stream should be set to
nullptr to mark the end.

ChatGPT suggests also in the constructor after calling get_next for
the first time. However, that would then be an element marked as end
whilst having a valid current element. In comparison, if the ++ operator
returns an end element, the last element will still be in current element.

Gotta check if the last one is still used in the iteration, if yes then we
need to change this, if the last element is not used and just there to check
if the end was reached, then we don't need it.