refactor(lookup): Use division instead of bitshift to divide by two
Took me quite some time to figure out why this bitshift is there and it
is just a division by two. Looked up if there is any performance gain
on this and according to the SO question linked below the compiler should
figure out the most efficient way to achieve the division.
Therfore, I changed it to a division as it makes the intent of the code more
clear.

https://stackoverflow.com/questions/6357038/is-multiplication-and-division-using-shift-operators-in-c-actually-faster