fix(lookuptable): Tangential position number is not the same as the nr of tangential bins
I used the number of tangential bins as the size for the LUT array,
however, also the min / max tangential number was half the detectors
per ring, which makes sense if you consider single detectors.

Looks like there is not a single bin per crystal pair, which is obvious
if you look at the number 344 for NSBINS, but where does that come from?