perf(event and tags): Use unique pointer for events and tags to avoid unnecessary copying
The unique pointer ensure that on each assignment to a member
the data behind the events or tags is not copied but only the
pointer.

Came up with the idea running perf and seeing 16% of the whole time
spent on 'rep_movs_alternative'. Avoinding copies only for event
and tag objects reduced the time spent in this to 10%.

More improvements should be possible elsewhere.