feat(image/sino): Have value type as template parameters
I plan to play around a bit with the data types to see what
impact it has on performance and memory consumption.
However, the current version could make problems due to the
following reason:

A sinogram from histograming holds int-like values, however,
a forward projected sinogram holds float-like values. Therefore,
one would theoretically need to variables in reconstruction if
one wants to make sure that rounding errors do not impact the
reconstruction.

Maybe  the benefit of using integers is not real, didn't check,
or the errors introduced are negligable. TBC